rca-agentic-backend-clean/requirements.txt
mcp
requests
python-dotenv
google-genai
google-adk>=2.0.0a0 # Installed via --pre
